requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
yfinance>=0.2.0
numpy>=1.24.0
schedule>=1.2.0
python-dotenv>=1.0.0
ib-insync>=0.9.0
statsmodels>=0.14.0
PyYAML>=6.0
scipy>=1.10.0